Various improvements
Added node tree searching. Added several tags to the list of those that are always self closing.

Fixed context switching tags like script, style, and block, which change the rules of the document when opened. The HTMLParser, when it finds such an opening tag, will not consider the tag to be closed until it finds a well-formed closing tag of the same name specifically.